requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
duckdb==0.9.2
exceptiongroup==1.2.0
executing==2.0.1
Faker==19.13.0
Flask==3.0.2
fonttools==4.49.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
packaging==23.2
pandas==1.5.3
pandasai==2.0.9
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
streamlit==1.32.1
tenacity==8.2.3
textblob==0.18.0.post0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wordcloud==1.9.3
